# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: APIs-Graduation
  version: 1.0.0
paths:
...
  /api/reset-password/send-code:
    post:
      tags:
        - 'Password Reset'
      summary: 'Request a password reset code'
      operationId: ee514344b04db5c54bb03ec253475e8f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  description: "Patient's email address"
                  type: string
                  format: email
                  example: patient@example.com
              type: object
      responses:
        200:
          description: 'Password reset code sent successfully'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'A password reset code has been sent to your email.'
                  status:
                    type: integer
                    example: 200
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Email is invalid or not registered.'
                  status:
                    type: integer
                    example: 422
                  errors:
                    type: object
                type: object
  /api/reset-password/resend-code:
    post:
      tags:
        - 'Password Reset'
      summary: 'Resend a password reset code'
      operationId: f246074e35b115782b954f0a1c5104d4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  description: "Patient's email address"
                  type: string
                  format: email
                  example: patient@example.com
              type: object
      responses:
        200:
          description: 'Password reset code resent successfully'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'A password reset code has been sent to your email.'
                  status:
                    type: integer
                    example: 200
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Email is invalid or not registered.'
                  status:
                    type: integer
                    example: 422
                  errors:
                    type: object
                type: object
  /api/reset-password/reset:
    post:
      tags:
        - 'Password Reset'
      summary: 'Reset patient password using code'
      operationId: 4afe1d5a9075541b8db8f5259f6dd923
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - code
                - password
                - password_confirmation
              properties:
                email:
                  description: "Patient's email address"
                  type: string
                  format: email
                  example: patient@example.com
                code:
                  description: '6-digit reset code'
                  type: string
                  example: '123456'
                password:
                  description: 'New password (min 8 characters, must include uppercase, lowercase, number, and special character)'
                  type: string
                  example: Password123!
                password_confirmation:
                  description: 'Confirmation of the new password'
                  type: string
                  example: Password123!
              type: object
      responses:
        200:
          description: 'Password reset successfully'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Password reset successfully.'
                  status:
                    type: integer
                    example: 200
                type: object
        422:
          description: 'Validation error or invalid/expired code'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The code is incorrect.'
                  status:
                    type: integer
                    example: 422
                  errors:
                    type: object
                    nullable: true
                type: object
...